rename some modules, add README
