+++ api/models/age_ranges.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from api.models.choices.age_range_types import AgeRangeType
+from api.models.choices.age_range_type import AgeRangeType
 from api.models.constants import DEFAULT_CHOICE_CHAR_FIELD_MAX_LENGTH
