+++ api/models/choices/README.md
@@ -1,0 +1,3 @@
+This package contains classes for [choices](https://docs.djangoproject.com/en/stable/ref/models/fields/#enumeration-types).
+
+Modules that contain only one class are named in singular, modules that contain several classes are named in plural.
